+++ deps.ts
@@ -1,1 +1,2 @@
-export { EventEmitter } from "https://deno.land/std@0.156.0/node/events.ts?s=EventEmitter";
+export { EventEmitter } from "https://deno.land/std@0.161.0/node/events.ts?s=EventEmitter";
+